Added cart page
This is the draft for the cart page. It features an order summary, the payment methods the user has on file, the promo code input box, and the options to confirm and cancel the order. NOT linked to the cart button found on the homepage, order success prototype still needs to be implemented on pressing "Confirm."